migrate more apis, cleanup and uncomment websocket
