+++ go.mod
@@ -7,10 +7,8 @@ require (
 	github.com/go-playground/validator/v10 v10.10.1
 	github.com/go-resty/resty/v2 v2.7.0
 	github.com/joho/godotenv v1.3.0
-	github.com/mitchellh/mapstructure v1.4.1
-	github.com/monaco-io/request v1.0.14
 	github.com/pquerna/otp v1.3.0
-	github.com/sacOO7/gowebsocket v0.0.0-20210515122958-9396f1a71e23
 	github.com/stretchr/testify v1.7.1
 	go.uber.org/zap v1.19.0
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
