--- order_type.go
@@ -1,8 +1,0 @@
-package goavanza
-
-type OrderType string
-
-const (
-	ORDER_TYPE_BUY  OrderType = "BUY"
-	ORDER_TYPE_SELL OrderType = "SELL"
-)
