--- websocket/client.go
@@ -1,8 +1,0 @@
-package websocket
-
-type AvanzaClient struct {
-}
-
-func New() *AvanzaClient {
-	return &AvanzaClient{}
-}
